requirements.txt
flask~=1.1.2
python_dotenv~=0.13.0
tweepy~=3.8.0
celery~=4.4.6
gunicorn~=20.0.4